+++ R/engine_inla.R
@@ -500,9 +500,17 @@ engine_inla <- function(x,
         # Messenger
         if(getOption('ibis.setupmessages')) myLog('[Estimation]','green','Engine setup.')
 
-        # Set number of threads via set.Options
-        INLA::inla.setOption(num.threads = getOption('ibis.nthread'),
-                             blas.num.threads = getOption('ibis.nthread'))
+        # MH: seems like blas.num.threads not available on windows?
+        if (.Platform$OS.type == "windows") {
+          # Set number of threads via set.Options
+          INLA::inla.setOption(num.threads = getOption('ibis.nthread'))
+        } else {
+          # Set number of threads via set.Options
+          INLA::inla.setOption(num.threads = getOption('ibis.nthread'),
+                               blas.num.threads = getOption('ibis.nthread'))
+        }
+
+
 
         # --- Prepare general inputs ---
         # Check whether spatial latent effects were added