+++ NEWS.md
@@ -7,6 +7,8 @@
 
 #### Minor improvements and bug fixes
 * Minor bug fixes in `scenario()` object, and MigClim and Kissmig wrappers.
+* Bug fix related to CRS classes of sp and sf
+* Bug fix related to blas.num.threads on Windows
 
 # ibis.iSDM 0.0.9